Write process ID as an unsigned int (with a cast). On Solaris, pid_t may be typedef'd as a long but the actual range is 32 bits at most.

+++ logsrvd/logsrvd.c
@@ -1680,7 +1680,7 @@ daemonize(bool nofork)
 	if (fp == NULL) {
 	    sudo_warn("%s", logsrvd_conf_pid_file());
 	} else {
-	    fprintf(fp, "%d\n", getpid());
+	    fprintf(fp, "%u\n", (unsigned int)getpid());
 	    fclose(fp);
 	}
     }
